requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-chroma>=0.1.0
chromadb>=0.4.0
pypdf>=3.17.0
pydantic>=2.0.0
python-dotenv>=1.0.0